app_03/requirements.txt
cheroot==8.3.0
jaroco.functools==3.0.1
more-intertools==8.4.0 
mysql-connector-python==8.0.20
pkg-resources==0.0.0
protobuf==3.12.2
six==1.15.0
web.py==0.51